+++ server/invoice_controller/helper/LeaseAgrm.php
@@ -1,8 +1,50 @@
 <?php 
+    Trait LeaseAgrmSql 
+    {
+        private function LeaseAgrmSql($extra_selects=[])
+        {
+            $selects = 
+            [
+                "`Rent_amount` AS `rent_amount`", 
+                "{$this->total_paid_amount} AS `paid_amount`", 
+                "{$this->total_invoice_amount} AS `total_amount`", 
+                "IFNULL((SELECT `leaseagrm_period`.`name` FROM `leaseagrm_period` WHERE `leaseagrm_period`.`id` = `leaseagrm_period_id`), 'months') AS `leaseagrm_period`"
+            ]; 
+
+            $selects = array_merge($selects, $extra_selects); 
+            return Query::SelectData("leaseagrm", $selects, ["id"=>$this->leaseagrm_id]); 
+        }
+
+        private function StartDateSql()
+        {
+            return 
+            "
+                SELECT MAX(`invoice_leaseagrm`.`end_date`) AS `start_date` FROM `invoice_leaseagrm`, `invoices` 
+                WHERE  
+                    `invoices`.`id` = `invoice_leaseagrm`.`invoice_id` AND 
+                    `invoice_leaseagrm`.`revenue_type_id` = '{$this->rent_id}' AND  
+                    `invoices`.`leaseagrm_id` = '{$this->leaseagrm_id}'; 
+            "; 
+        }
+
+        private function RentInformationSql()
+        {
+            return 
+            "                    
+                SELECT * FROM `invoice_leaseagrm` 
+                WHERE 
+                    `invoice_id` IN (SELECT `id` FROM `invoices` WHERE `leaseagrm_id` = '{$this->leaseagrm_id}') 
+                    AND `revenue_type_id` = '{$this->rent_id}' 
+                ORDER BY `start_date` ASC; 
+            "; 
+        }
+    }
+
     Trait LeaseAgrm
     {
+        use LeaseAgrmSql; 
         private $leaseagrm_id, $rent_id, $total_paid_amount, $total_invoice_amount, $start_lease; 
-        private $date_format = "Y-m-d"; 
+        private $date_format = "Y-m-d", $test_mode; 
         function __construct($leaseagrm_id)
         {
             $helper_path = __DIR__ . "/../../helper"; 
@@ -13,40 +55,24 @@
             $this->rent_id = OverviewQueries\Invoices::RentId(); 
             $this->total_paid_amount = OverviewQueries\LeaseAgrm::TotalPaidAmountQuery(); 
             $this->total_invoice_amount = OverviewQueries\LeaseAgrm::$TotalInvoiceAmountQuery; 
+
+            $this->test_mode = CurrentEnvironment::TestMode(); 
         }
 
-        private function LeaseAgrm()
+        private function LeaseAgrm($leaseagrm=null, $start_date_data=null, $rent_information=null)
         {
-            $selects = 
-            [
-                "`Rent_amount` AS `rent_amount`", 
-                "{$this->total_paid_amount} AS `paid_amount`", 
-                "{$this->total_invoice_amount} AS `total_amount`", 
-                "IFNULL((SELECT `leaseagrm_period`.`name` FROM `leaseagrm_period` WHERE `leaseagrm_period`.`id` = `leaseagrm_period_id`), 'months') AS `leaseagrm_period`"
-            ]; 
-        
-            $sql = Query::SelectData("leaseagrm", $selects, ["id"=>$this->leaseagrm_id]); 
-            $leaseagrm = ConnectSqlite::Query($sql); 
-            $leaseagrm = $leaseagrm[0]; 
+            $leaseagrm = $leaseagrm?? ConnectSqlite::Query($this->LeaseAgrmSql())[0]; 
             $leaseagrm["difference"] = floatval($leaseagrm["total_amount"]) - floatval($leaseagrm["paid_amount"]); 
-            $start_date = $this->StartDate(); 
+            $start_date = $this->StartDate($start_date_data); 
             $leaseagrm["start_date"] = $start_date; 
             $leaseagrm["end_date"] = $this->EndDate($start_date); 
-            $leaseagrm["rent_information"] = $this->RentInformation(); 
+            $leaseagrm["rent_information"] = $this->RentInformation($rent_information); 
             return $leaseagrm; 
         }
 
-        private function StartDate()
+        private function StartDate($data=null)
         {
-            $sql = 
-            "
-                SELECT MAX(`invoice_leaseagrm`.`end_date`) AS `start_date` FROM `invoice_leaseagrm`, `invoices` 
-                WHERE  
-                    `invoices`.`id` = `invoice_leaseagrm`.`invoice_id` AND 
-                    `invoice_leaseagrm`.`revenue_type_id` = '{$this->rent_id}' AND  
-                    `invoices`.`leaseagrm_id` = '{$this->leaseagrm_id}'
-            "; 
-            $data = ConnectSqlite::Query($sql); 
+            $data = $data?? ConnectSqlite::Query($this->StartDateSql()); 
             if($data[0]["start_date"])
             {
                 return (new DateTime($data[0]["start_date"]))->modify("+1 day")->format($this->date_format); 
@@ -72,17 +98,9 @@
             }
         }
 
-        private function RentInformation()
+        private function RentInformation($data=null)
         {
-            $sql = 
-            "                    
-                SELECT * FROM `invoice_leaseagrm` 
-                WHERE 
-                    `invoice_id` IN (SELECT `id` FROM `invoices` WHERE `leaseagrm_id` = '{$this->leaseagrm_id}') 
-                    AND `revenue_type_id` = '{$this->rent_id}' 
-                ORDER BY `start_date` ASC 
-            "; 
-            $data = ConnectSqlite::Query($sql); 
+            $data = $data?? ConnectSqlite::Query($this->RentInformationSql()); 
             if(!count($data))
             {
                 return [["start_date"=>(new DateTime($this->start_lease))->format($this->date_format), "end_date"=> null]];
